Implicitly pass the postgres url to manager
