+++ operator/pkg/controllers/hubofhubs/globalhub_postgres.go
@@ -2,6 +2,7 @@ package hubofhubs
 
 import (
 	"context"
+	"encoding/base64"
 	"fmt"
 
 	postgresv1beta1 "github.com/crunchydata/postgres-operator/pkg/apis/postgres-operator.crunchydata.com/v1beta1"
@@ -10,17 +11,36 @@ import (
 	"k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/discovery"
+	"k8s.io/client-go/discovery/cached/memory"
+	"k8s.io/client-go/restmapper"
 
 	globalhubv1alpha4 "github.com/stolostron/multicluster-global-hub/operator/apis/v1alpha4"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/config"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/constants"
+	operatorconstants "github.com/stolostron/multicluster-global-hub/operator/pkg/constants"
+	"github.com/stolostron/multicluster-global-hub/operator/pkg/deployer"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/postgres"
+	"github.com/stolostron/multicluster-global-hub/operator/pkg/renderer"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/utils"
 )
 
 const (
-	secretNilErrorMsg = "postgres secret %s is nil"
+	secretNilErrorMsg        = "postgres secret %s is nil"
+	postgresAdminUsername    = "postgres"
+	postgresReadonlyUsername = "global-hub-readonly-user" // #nosec G101
+	postgresCA               = "multicluster-global-hub-postgres-ca"
 )
+
+type postgresCredential struct {
+	postgresAdminUsername        string
+	postgresAdminUserPassword    string
+	postgresReadonlyUsername     string
+	postgresReadonlyUserPassword string
+}
+
+var partialPostgresURI = "@multicluster-global-hub-postgres." +
+	config.GetDefaultNamespace() + ".svc:5432/hoh?sslmode=verify-ca"
 
 // EnsureCrunchyPostgresSubscription verifies resources needed for Crunchy Postgres are created
 func (r *MulticlusterGlobalHubReconciler) EnsureCrunchyPostgresSubscription(ctx context.Context,
@@ -67,17 +87,13 @@ func (r *MulticlusterGlobalHubReconciler) EnsureCrunchyPostgres(ctx context.Cont
 		Name:      postgres.PostgresName,
 		Namespace: config.GetDefaultNamespace(),
 	}, postgresCluster)
-	if err != nil {
-		if errors.IsNotFound(err) {
-			err = r.Client.Create(ctx, postgres.NewPostgres(postgres.PostgresName, config.GetDefaultNamespace()))
-			if err != nil {
-				return err
-			}
-			return nil
+	if err != nil && errors.IsNotFound(err) {
+		e := r.Client.Create(ctx, postgres.NewPostgres(postgres.PostgresName, config.GetDefaultNamespace()))
+		if e != nil {
+			return e
 		}
-		return err
 	}
-	return nil
+	return err
 }
 
 // WaitForPostgresReady waits for postgres to be ready and returns a postgres connection
@@ -145,4 +161,116 @@ func (r *MulticlusterGlobalHubReconciler) GeneratePGConnectionFromGHStorageSecre
 		ReadonlyUserDatabaseURI: string(pgSecret.Data["database_uri_with_readonlyuser"]),
 		CACert:                  pgSecret.Data["ca.crt"],
 	}, nil
+}
+
+func (r *MulticlusterGlobalHubReconciler) initPostgresByStatefulset(ctx context.Context,
+	mgh *globalhubv1alpha4.MulticlusterGlobalHub,
+) error {
+	// install the postgres statefulset only
+	credential, err := getPostgresCredential(ctx, mgh, r)
+	if err != nil {
+		return err
+	}
+	imagePullPolicy := corev1.PullAlways
+	if mgh.Spec.ImagePullPolicy != "" {
+		imagePullPolicy = mgh.Spec.ImagePullPolicy
+	}
+
+	// get the postgres objects
+	postgresRenderer, postgresDeployer := renderer.NewHoHRenderer(fs), deployer.NewHoHDeployer(r.Client)
+	postgresObjects, err := postgresRenderer.Render("manifests/postgres", "",
+		func(profile string) (interface{}, error) {
+			return struct {
+				Namespace                    string
+				PostgresImage                string
+				StorageSize                  string
+				ImagePullSecret              string
+				ImagePullPolicy              string
+				NodeSelector                 map[string]string
+				Tolerations                  []corev1.Toleration
+				PostgresAdminUserPassword    string
+				PostgresReadonlyUsername     string
+				PostgresReadonlyUserPassword string
+				PostgresCACert               string
+				StorageClass                 string
+			}{
+				Namespace:                    config.GetDefaultNamespace(),
+				PostgresImage:                config.GetImage(config.PostgresImageKey),
+				ImagePullSecret:              mgh.Spec.ImagePullSecret,
+				ImagePullPolicy:              string(imagePullPolicy),
+				NodeSelector:                 mgh.Spec.NodeSelector,
+				Tolerations:                  mgh.Spec.Tolerations,
+				StorageSize:                  config.GetPostgresStorageSize(mgh),
+				PostgresAdminUserPassword:    credential.postgresAdminUserPassword,
+				PostgresReadonlyUsername:     credential.postgresReadonlyUsername,
+				PostgresReadonlyUserPassword: credential.postgresReadonlyUserPassword,
+				PostgresCACert:               base64.StdEncoding.EncodeToString(r.MiddlewareConfig.PgConnection.CACert),
+				StorageClass:                 mgh.Spec.DataLayer.StorageClass,
+			}, nil
+		})
+	if err != nil {
+		return fmt.Errorf("failed to render postgres manifests: %w", err)
+	}
+
+	// create restmapper for deployer to find GVR
+	dc, err := discovery.NewDiscoveryClientForConfig(r.Manager.GetConfig())
+	if err != nil {
+		return err
+	}
+	mapper := restmapper.NewDeferredDiscoveryRESTMapper(memory.NewMemCacheClient(dc))
+
+	if err = r.manipulateObj(ctx, postgresDeployer, mapper, postgresObjects, mgh, r.Log); err != nil {
+		return fmt.Errorf("failed to create/update postgres objects: %w", err)
+	}
+
+	ca, err := getPostgresCA(ctx, mgh, r)
+	if err != nil {
+		return err
+	}
+	r.MiddlewareConfig.PgConnection = &postgres.PostgresConnection{
+		SuperuserDatabaseURI: "postgresql://" + credential.postgresAdminUsername + ":" +
+			credential.postgresAdminUserPassword + partialPostgresURI,
+		ReadonlyUserDatabaseURI: "postgresql://" + credential.postgresReadonlyUsername + ":" +
+			credential.postgresReadonlyUserPassword + partialPostgresURI,
+		CACert: []byte(ca),
+	}
+	return nil
+}
+
+func getPostgresCredential(ctx context.Context, mgh *globalhubv1alpha4.MulticlusterGlobalHub,
+	r *MulticlusterGlobalHubReconciler,
+) (*postgresCredential, error) {
+	postgres := &corev1.Secret{}
+	if err := r.Client.Get(ctx, types.NamespacedName{
+		Name:      operatorconstants.GHBuiltInStorageSecretName,
+		Namespace: mgh.Namespace,
+	}, postgres); err != nil && errors.IsNotFound(err) {
+		return &postgresCredential{
+			postgresAdminUsername:        postgresAdminUsername,
+			postgresAdminUserPassword:    generatePassword(16),
+			postgresReadonlyUsername:     postgresReadonlyUsername,
+			postgresReadonlyUserPassword: generatePassword(16),
+		}, nil
+	} else if err != nil {
+		return nil, err
+	}
+	return &postgresCredential{
+		postgresAdminUsername:        postgresAdminUsername,
+		postgresAdminUserPassword:    string(postgres.Data["database-admin-password"]),
+		postgresReadonlyUsername:     string(postgres.Data["database-readonly-user"]),
+		postgresReadonlyUserPassword: string(postgres.Data["database-readonly-password"]),
+	}, nil
+}
+
+func getPostgresCA(ctx context.Context,
+	mgh *globalhubv1alpha4.MulticlusterGlobalHub, r *MulticlusterGlobalHubReconciler,
+) (string, error) {
+	ca := &corev1.ConfigMap{}
+	if err := r.Client.Get(ctx, types.NamespacedName{
+		Name:      postgresCA,
+		Namespace: mgh.Namespace,
+	}, ca); err != nil {
+		return "", err
+	}
+	return ca.Data["service-ca.crt"], nil
 }
