+++ operator/pkg/controllers/hubofhubs/manifests/manager/postgres-credential-secret.yaml
@@ -1,10 +1,11 @@
 apiVersion: v1
 kind: Secret
 metadata:
-  name: postgres-certs-secret
+  name: postgres-credential-secret
   namespace: {{.Namespace}}
   labels:
     name: multicluster-global-hub-manager
 type: Opaque
 data:
   "ca.crt": "{{.PostgresCACert}}"
+  "database-url": "{{.DatabaseURL}}"
