+++ operator/pkg/controllers/hubofhubs/manifests/postgres/secret.yaml
@@ -12,3 +12,4 @@ stringData:
   database-admin-password: "{{.PostgresAdminUserPassword}}"
   database-readonly-user: "{{.PostgresReadonlyUsername}}"
   database-readonly-password: "{{.PostgresReadonlyUserPassword}}"
+  "ca.crt": "{{.PostgresCACert}}"
